requirements.txt
Flask
opencv-python
opencv-python-headless
numpy
picamera[array]